+++ sys/build.sh
@@ -4,6 +4,7 @@ MAKE=make
 gmake --help >/dev/null 2>&1
 [ $? = 0 ] && MAKE=gmake
 [ -z "${MAKE_JOBS}" ] && MAKE_JOBS=12
+[ -z "${CERTID}" ] && CERTID=org.radare.radare2
 
 # find root
 cd `dirname $PWD/$0` ; cd ..
@@ -59,7 +60,7 @@ echo "export USE_R2_CAPSTONE=$USE_R2_CAPSTONE"
 echo
 # Set USE_R2_CAPSTONE env var to ignore syscapstone check
 if [ -z "${USE_R2_CAPSTONE}" ]; then
-	pkg-config --atleast-version=3.0 capstone 2>/dev/null
+	pkg-config --atleast-version=4.0 capstone 2>/dev/null
 	if [ $? = 0 ]; then
 		echo '#include <capstone.h>' > .a.c
 		echo 'int main() {return 0;}' >> .a.c
@@ -81,7 +82,10 @@ if [ -d shlr/capstone/.git ]; then
 ( cd shlr/capstone ; git clean -xdf )
 fi
 [ "`uname`" = Linux ] && export LDFLAGS="-Wl,--as-needed ${LDFLAGS}"
-rm -f plugins.cfg
+if [ -z "${KEEP_PLUGINS_CFG}" ]; then
+	rm -f plugins.cfg
+fi
+unset DEPS
 ./configure ${CFGARG} --prefix=${PREFIX} || exit 1
 ${MAKE} -s -j${MAKE_JOBS} MAKE_JOBS=${MAKE_JOBS} || exit 1
 if [ "`uname`" = Darwin ]; then
